# openapi.yaml
openapi: 3.1.0
info:
  title: Preloop AI API
  description: REST API for Preloop AI issue tracking management
  version: 0.1.0
paths:
...
  /api/v1/issues/{issue_id}:
    get:
      tags:
      - Issues
      summary: Get Issue
      description: Get details of a specific issue using its external ID.
      operationId: get_issue_api_v1_issues__issue_id__get
      security:
      - bearerAuth: []
      parameters:
      - name: issue_id
        in: path
        required: true
        schema:
          type: string
          title: Issue Id
      - name: organization
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Organization
      - name: project
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Project
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Issues
      summary: Update Issue
      description: Update an existing issue using its internal ID or external key.
      operationId: update_issue_api_v1_issues__issue_id__put
      security:
      - bearerAuth: []
      parameters:
      - name: issue_id
        in: path
        required: true
        schema:
          type: string
          title: Issue Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IssueResponse:
      properties:
        title:
          type: string
          title: Title
          description: Issue title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: Issue description
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
          description: Issue priority (Low, Medium, High)
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
          description: Issue status
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
          description: Issue assignee
        labels:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Labels
          description: Issue labels
        meta_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta Data
          description: Additional issue metadata
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
          description: Internal Preloop AI database ID (UUID)
        external_id:
          type: string
          title: External Id
          description: Issue ID in the original tracker
        key:
          type: string
          title: Key
          description: Human-readable issue key (e.g., PROJ-123)
        organization:
          type: string
          title: Organization
          description: Organization name
        project:
          type: string
          title: Project
          description: Project name
        project_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Id
          description: Project ID
        project_identifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Identifier
          description: Project identifier/slug for API calls
        url:
          type: string
          title: Url
          description: URL to the issue in the original tracker
        created_at:
          type: string
          title: Created At
          description: Creation timestamp
        updated_at:
          type: string
          title: Updated At
          description: Last update timestamp
        comments:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Comments
          description: List of comments on the issue
          default: []
        score:
          anyOf:
          - type: number
          - type: 'null'
          title: Score
          description: Similarity score for search results (if applicable)
      type: object
      required:
      - title
      - id
      - external_id
      - key
      - organization
      - project
      - project_id
      - url
      - created_at
      - updated_at
      title: IssueResponse
      description: Response model for issue data.
    IssueUpdate:
      properties:
        organization:
          anyOf:
          - type: string
          - type: 'null'
          title: Organization
          description: Organization name
        project:
          anyOf:
          - type: string
          - type: 'null'
          title: Project
          description: Project name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
          description: New issue title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          description: New issue description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
          description: New issue status
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
          description: New issue priority
        assignee:
          anyOf:
          - type: string
          - type: 'null'
          title: Assignee
          description: New issue assignee
        labels:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Labels
          description: New issue labels
        meta_data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta Data
          description: Additional issue metadata
      type: object
      title: IssueUpdate
      description: Model for updating an issue.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication